Create nse-service-finder.sh
This Bash script accepts one (1) argument from the command line interface: the user's desired NMAP scripting engine (NSE) script. With the argument, the tool parses all NSE scripts applicable to the user's service and returns the name of the NSE script, as well as its syntax and arguments. It is meant to save the user time and enable them to more quickly harness the power of NSE scripts in NMAP.